+++ common/src/main/java/com/faboslav/structurify/common/util/StructurePieceUtil.java
@@ -1,0 +1,31 @@
+package com.faboslav.structurify.common.util;
+
+import com.faboslav.structurify.common.mixin.structure.pools.SinglePoolElementAccessor;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.level.levelgen.structure.PoolElementStructurePiece;
+import net.minecraft.world.level.levelgen.structure.StructurePiece;
+import net.minecraft.world.level.levelgen.structure.pools.SinglePoolElement;
+import net.minecraft.world.level.levelgen.structure.pools.StructurePoolElement;
+
+public class StructurePieceUtil
+{
+	public static String getStructurePieceName(StructurePiece structurePiece) {
+		String pieceName = "";
+
+		if (structurePiece instanceof PoolElementStructurePiece poolPiece) {
+			StructurePoolElement element = poolPiece.getElement();
+
+			if (element instanceof SinglePoolElement single) {
+				ResourceLocation template = ((SinglePoolElementAccessor) single).getTemplate().left().orElse(null);
+
+				if (template != null) {
+					pieceName = template.toString();
+				}
+			} else {
+				pieceName = "";
+			}
+		}
+
+		return pieceName;
+	}
+}